requirements.txt
requests
python-dotenv
setuptools
PyQt5
pyinstaller
configparser 
hearthstone
aniso8601